Make master buildable with latest bindings-GLFW.
This change is meant to facilitate a healthy ecosystem w.r.t. Stackage. Although
the latest version on Hackage has the proper version constraints to build via
cabal, Stackage nightly snapshots will download the newest versions of all
packages to see if compatibility is up-to-date. To conform with this
restriction, GLFW-b should also build against 3.2.1 even though not all of the
features from 3.2.1 are implemented.

Some side effects:
- Moved StandardCursorShape into genC.hs
- Some small formatting
- Added bindings-GLFW-3.2.1.0 to extra-deps in stack.yaml
- Bumped cabal version
- Updated bindings-GLFW version constraints in cabal file